+++ build.bfg
@@ -1,0 +1,8 @@
+# -*- python -*-
+
+main = executable(
+    'bullet',
+    files=find_files('bullet', '*.cc', extra='*.h'),
+    compile_options(
+      ['std=c++1z', '-Wall', '-Werror', '-Wextra', '-pedantic'],
+    link_options=['-lglog'])
