--- bullet/include/bullet/util.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <string>
-
-namespace bullet {
-  class Util {
-    public:
-      void check_dir(const std::string& path);
-  };
-}// namespace bullet
